# Optimize sort_by_key (SortFaces/SortVerts) 