Remap boxes separately

MacOS M1/TBB: 1485ms -> 1385ms