@pca006132 2023-07-16
interesting, I guess these recent optimizations means that sort_by_key and sort using zip iterators are very slow? I originally thought that the performance should be fine.
will have a look more carefully later today