@stephomi 2023-07-16
Nah sort_by_key is fine (actually I could have reverted to sort_by_key in the previous PR, it would give similar performance).

The only thing that matters is the size of the object to move (both keys or values), here by sorting boxes we are moving big objects for nothing.
If you plan on sync-sorting multiple array it’s best to to perform the sort on the minimal amount of memory and re-index the rest once.