@pca006132 2023-07-16
For the build error, I think we hit the #460. The safer way I can now think of is to use clang. Try to add the following commit:
```diff
diff --git a/flake.nix b/flake.nix
index 04ffc90..4c37ab3 100644
--- a/flake.nix
+++ b/flake.nix
@@ -16,7 +16,7 @@
             , doCheck ? true
             , build-tools ? [ ]
             , ...
-            }: pkgs.stdenv.mkDerivation {
+            }: pkgs.llvmPackages_15.stdenv.mkDerivation {
               inherit doCheck;
               pname =
                 if cuda-support then
```

> Not sure if I'm breaking the whole CUDA thing though, if I don't get the pointer first, it just takes forever (unless I capture by reference [&] instead of value [=])

it seems that this is no longer relevant?

> MacOS M1/TBB: 1485ms -> 1385ms

Wondering what benchmark are you using? Wonder if you can share them to the unit tests? This can help us track performance later.