@stephomi 2023-07-16
> it seems that this is no longer relevant?

Yep, removed the comment to avoid confusion

> Wondering what benchmark are you using? Wonder if you can share them to the unit tests? This can help us track performance later.

A simple sphere/cylinder subtraction.

<img width="551" alt="Capture d’écran 2023-07-16 à 20 02 37" src="https://github.com/elalish/manifold/assets/4750281/cd3a01f7-5fd5-48e2-8ac7-cf638d555404">

I'm always benching the part that I'm editing, since there's a bit of variance by looking at the whole thing.
For the PR I'm giving timing difference for the whole operation to give an idea of the gain (minor here, but the idea is also to give insight on what might be improved)

The part in question:
|             | Master        | PR           |
|-------------| ------------- | ------------ |
| SortVerts 1 | 22 ms         | 17 ms        |
| SortVerts 2 | 4 ms          | 3  ms        |
| SortVerts 3 | 20 ms         | 14 ms        |
| SortFaces 1 | 100 ms        | 60 ms        |
| SortFaces 2 | 15 ms         | 10 ms        |
| SortFaces 3 | 97 ms         | 56 ms        |